Renaming and reorganization of interfaces
