+++ src/console/interfaces/__init__.py
@@ -1,0 +1,1 @@
+"""Interfaces init file."""
